+++ src/lib.rs
@@ -65,7 +65,7 @@ it in your crate, add it as a dependency inside [*Cargo.toml*]:
 
 ```toml
 [dependencies]
-fixed = "0.1.1"
+fixed = "0.1.2"
 ```
 
 You also need to declare it by adding this to your crate root (usually
@@ -114,7 +114,7 @@ additional terms or conditions.
 */
 #![no_std]
 #![warn(missing_docs)]
-#![doc(html_root_url = "https://docs.rs/fixed/0.1.1")]
+#![doc(html_root_url = "https://docs.rs/fixed/0.1.2")]
 #![doc(test(attr(deny(warnings))))]
 #![cfg_attr(nightly_repr_transparent, feature(repr_transparent))]