+++ RELEASES.md
@@ -10,6 +10,8 @@ Version 0.1.2 (unreleased)
 
   * The crate can now be used without the standard library `std`.
   * New methods `from_f32` and `from_f64` were added.
+  * New methods `is_positive` and `is_negative` were added to signed
+    fixed-point numbers.
 
 Version 0.1.1 (2018-08-11)
 ==========================