+++ RELEASES.md
@@ -5,7 +5,7 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
-Version 0.1.2 (unreleased)
+Version 0.1.2 (2018-08-15)
 ==========================
 
   * The crate can now be used without the standard library `std`.